plotter: ironing out units
Currently, Gqrx shows 0dB for a 1 unit peak sine wave, so it is
using dBFS (full scale) with respect to peak (not RMS).

Keep the current behavior, naming it dBFS mode.

For dBmW mode, interpret the signal as V into a 50 ohm load.
A 1V peak sine wave into a 50 ohm load is 10 mW, or 10 dBm.

Signed-off-by: Jeff Long <willcode4@gmail.com>